feat: 항상 TOP100에 대해서만 저장
